refactor(auth): move load_auth_from_storage body to _auth/tokens.py (Wave 3a / Task 2.1)
Wave 3a of the session-decoupling plan
(docs/session-decoupling-plan-2026-05-26.md). Closes Stage 6 of the
architecture-fix-plan: the auth.py facade's body-bearing functions move into
the _auth/ subpackage; the facade becomes a stable re-export point.

Changes:

- _auth/tokens.py
  - Import psidts_recovery (needed by the function body's recovery branch).
  - Add `load_auth_from_storage(path)` with the full body cut from auth.py.
    Internal cross-module references go through the explicit `_auth_cookies`
    alias (same pattern AuthTokens uses) instead of the auth.py-level
    `_load_storage_state` / `extract_cookies_from_storage` shortcuts.

- auth.py
  - Replace the multi-line function with `load_auth_from_storage = _auth_tokens.load_auth_from_storage`.
  - Add `from ._auth import tokens as _auth_tokens` module alias.
  - Drop `from pathlib import Path` (no longer used after the body moved).

- tests/unit/test_auth_storage.py
  - Add `test_load_auth_from_storage_lives_in_private_module` — identity pin
    so a future drift (someone re-introducing a distinct body on auth.py)
    fails at PR time.
  - Existing `notebooklm.auth.load_auth_from_storage` test imports are
    unchanged: the re-export keeps the public surface stable.

Verification:
- 5829 passed, 10 skipped, 0 failed on `uv run pytest tests/unit tests/_lint`.
- ruff format + check + mypy all green.
- test_public_shims.py pins still hold (the public-surface re-export is
  identity-preserving).

This PR is **parallel** with Wave 2 (#1065) — zero file overlap — so it
targets `main` directly. Bot review can proceed in parallel.
